+++ Changes.md
@@ -25,6 +25,7 @@ API
 - ViewportGadget :
   - Added const overloads of `getCenterOfInterest()` and `getMaxPlanarZoom()`. The non-const versions will be removed in a future version.
   - Added `setTumblingEnabled()` and `getTumblingEnabled()` methods.
+  - Added `setDollyingEnabled()` and `getDollyingEnabled()` methods.
 
 0.61.5.0 (relative to 0.61.4.0)
 ========